--- src/components/TodoItem.js
@@ -1,30 +1,0 @@
-import { Component } from 'react';
-import styles from './TodoItem.module.scss';
-
-class TodoItem extends Component {
-  render() {
-    const completedStyle = {
-      fontStyle: 'italic',
-      color: '#595959',
-      opacity: 0.4,
-      textDecoration: 'line-through',
-    };
-    const { id, title, completed } = this.props.todo;
-    return (
-      <li className={styles.item}>
-        <input
-          type="checkbox"
-          className={styles.checkbox}
-          checked={completed}
-          onChange={() => this.props.handleChangeAttr(id)}
-        />
-        <button onClick={() => this.props.deleteTodosAttr(id)}>Delete</button>
-        <span style={completed ? completedStyle : null}>
-          {title}
-        </span>
-      </li>
-    );
-  }
-}
-
-export default TodoItem;
