+++ src/index.js
@@ -1,16 +1,15 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 
-// class-based component
-import TodoContainer from './components/TodoContainer';
+import TodoContainer from "./functionBased/components/TodoContainer"
 
-// stylesheet
-import './App.css';
+//stylesheet
+import "./functionBased/App.css"
 
 const root = document.getElementById('root');
 const container = createRoot(root);
 container.render(
-  <React.StrictMode>
-    <TodoContainer />
-  </React.StrictMode>,
+   <React.StrictMode>
+      <TodoContainer />
+   </React.StrictMode>,
 );
